decorate fast-memoize
