--- src/app/common/domain/smooth-setter.ts
@@ -1,41 +1,0 @@
-import { interval, map, startWith, Subject, take, takeUntil } from 'rxjs';
-
-export class SmoothSetter<T> {
-
-  unsubscribe$ = new Subject<void>();
-
-  constructor(public value: T,
-              private intervalTime: number,
-              private steps: number,
-              private setCallback: (lerp: number, newValue: T, oldValue: T) => T,
-              private doneCallback: () => void = () => void 0) {
-    // todo: setup default setCallback based on T type
-  }
-
-  set(newValue: T) {
-    this.unsubscribe$.next();
-
-    interval(this.intervalTime)
-      .pipe(
-        take(this.steps - 1),
-        map(i => i + 2),
-        startWith(1),
-        takeUntil(this.unsubscribe$))
-      .subscribe(index => {
-        this.value = this.setCallback(index / this.steps, newValue, this.value);
-        this.doneCallback();
-      });
-  }
-
-  destroy() {
-    this.unsubscribe$.next();
-    this.unsubscribe$.complete();
-  }
-
-}
-
-// todo: requestAnimationFrame to improve visual performance
-// window.requestAnimationFrame(() => {
-//   let val = `translate3D(${posX}px, ${posY}px, 0px) scale(${scale})`
-//   node.style.transform = val
-// })
