+++ src/app/pages/commnet-planner/models/connection-graph.ts
@@ -1,12 +1,11 @@
-import { environment } from '../../../environments/environment';
-import { ProbeControlPoint } from '../../common/domain/antenna';
+import { ProbeControlPoint } from './antenna';
 import {
   AntennaSignal,
   CanCommunicate,
-} from '../../common/domain/antenna-signal';
-import { NodeGraph } from '../../common/domain/graph-data-structure';
-import { Craft } from '../../common/domain/space-objects/craft';
-import { Planetoid } from '../../common/domain/space-objects/planetoid';
+} from './antenna-signal';
+import { NodeGraph } from '../../../common/domain/graph-data-structure';
+import { Craft } from '../../../common/domain/space-objects/craft';
+import { Planetoid } from '../../../common/domain/space-objects/planetoid';
 
 export class ConnectionGraph {
 
@@ -47,7 +46,7 @@ export class ConnectionGraph {
         }
 
         let otherNode = signal.nodes.find(n => n !== c);
-        let isGuidanceControlled = this.hostNodeCanControl(signal, otherNode, c);
+        let isGuidanceControlled = this.hostNodeCanControl(signal, otherNode);
 
         if (isGuidanceControlled) {
           this.hasControlCraft.add(c);
@@ -59,9 +58,7 @@ export class ConnectionGraph {
         if (isOtherNodeInNetwork) {
 
           let hasConnectionToBase = controlStationsAndCores
-              .some(cb => this.graph
-                  .shortestPath(otherNode.label, cb.label)
-                  .length);
+            .some(cb => this.graph.shortestPath(otherNode.label, cb.label).length);
           if (hasConnectionToBase) {
             this.hasControlCraft.add(c);
             break;
@@ -72,13 +69,12 @@ export class ConnectionGraph {
   }
 
   private hostNodeCanControl(signal: AntennaSignal,
-                             hostNode: CanCommunicate,
-                             clientNode: CanCommunicate): boolean {
+                             hostNode: CanCommunicate): boolean {
     let hasDsnCapability = hostNode instanceof Planetoid
       && hostNode.communication?.antennae.length;
 
     let hostCanRelayConnectionToClient = hostNode.communication.bestRemoteGuidanceCapability()
-      && signal.getHostToClientSignalStrength(hostNode, clientNode);
+      && signal.getHostToClientSignalStrength(hostNode); // signal is of this craft; the other node has to be the craft
 
     let hasControlConnection = hasDsnCapability || hostCanRelayConnectionToClient;
     return !!hasControlConnection;
